requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
decorator==4.4.1
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
mock==3.0.5
model-bakery==1.0.2
more-itertools==7.2.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.5
pytest==5.2.4
pytest-cov==2.8.1
pytest-django==3.7.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
coveralls==1.8.2